genio-mvp/backend/pyproject.toml
[tool.poetry]
name = "genio-backend"
version = "0.1.0"
description = "Genio Knowledge OS - Stream MVP Backend"
authors = ["Genio Team <team@genio.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
qdrant-client = "^1.7.0"
litellm = "^1.0.0"
feedparser = "^6.0.11"
trafilatura = "^1.6.0"
requests = "^2.31.0"
aiohttp = "^3.9.1"
httpx = "^0.26.0"
sendgrid = "^6.11.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
tenacity = "^8.2.3"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
